python_services/requirements.txt
flask
flask-cors
opencv-python-headless
mediapipe
scikit-learn
pypdf
numpy<2
fer
tensorflow
librosa
openai-whisper

deepface
tf-keras
